# docs/swagger.yaml
basePath: /
definitions:
  auth.forgotPasswordForm:
    properties:
      email:
        type: string
    type: object
  auth.loginForm:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  auth.registerForm:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
info:
  contact: {}
  description: Proman Backend API
  title: Proman Backend
paths:
  /api/forgot-password:
    post:
      consumes:
      - application/json
      operationId: forgot-password
      parameters:
      - description: forgot password json
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.forgotPasswordForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: ForgotPassword
      tags:
      - Auth
  /api/login:
    post:
      consumes:
      - application/json
      operationId: login
      parameters:
      - description: login json
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.loginForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Login
      tags:
      - Auth
...
  /api/register:
    post:
      consumes:
      - application/json
      operationId: register
      parameters:
      - description: register json
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.registerForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Register
      tags:
      - Auth
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"